+++ lib/nuget-parser/csproj-parser.ts
@@ -6,6 +6,7 @@ import * as parseXML from 'xml2js';
 import * as _ from 'lodash';
 import * as debugModule from 'debug';
 import { TargetFramework } from './types';
+import { toReadableFramework } from './framework';
 const debug = debugModule('snyk');
 
 export async function getTargetFrameworksFromProjFile(rootDir: string): Promise<TargetFramework | undefined> {
@@ -43,25 +44,6 @@ export async function getTargetFrameworksFromProjFile(rootDir: string): Promise<
     debug('.csproj file not found in ' + rootDir + '.');
     resolve();
   });
-}
-
-function toReadableFramework(targetFramework: string): TargetFramework | undefined {
-  const typeMapping = {
-    net: '.NETFramework',
-    netcoreapp: '.NETCore',
-    netstandard: '.NETStandard',
-    v: '.NETFramework',
-  };
-
-  for (const type in typeMapping) {
-    if (new RegExp(type + /\d.?\d(.?\d)?$/.source).test(targetFramework)) {
-      return {
-        framework: typeMapping[type],
-        original: targetFramework,
-        version: targetFramework.split(type)[1],
-      };
-    }
-  }
 }
 
 function findFile(rootDir, filter) {
